Make crash tests possible in release
Crash tests are made possible in release via
throwing an exception rather then asserting.
Using preprocessor, we can check that the code
threw or asserted based on the debug build.

This way, we keep asserts in debug builds, but
we also get the relevant crash message when crashing
in the same way in release.