Reworked memory + loop tracking and input layout
This commit adds support for fetching the input sources for any result
ID.
This is intended to be used to create maps from output to input sources
in a robust manner, which in turn enables advanced tracking of values
through long dispatch chains.
Also adds support for tracking structured loops in a clean manner so we
can detect infinite loops during testing.

Change-Id: I177da44ff2104b7d2f4a3d290f69672939f170c8